Modified to work in a local network
## client/scripts/network.js
When getting the endpoint string, replace the port (i.e. :5000) with :3000. I'm doing this so that it is possible to run the server `node index.js` and run `npm serve` for the client folder on separate ports and still work.

## server/index.js
Replaced all mentions of `this._rooms[peer.ip]` and `this._rooms[sender.ip]` with `this._rooms.local` to let everyone on the network see each other.